Changed AlleleDepth.getDepthForAllele() to return int instead of byte.  Also added new getDepth() method.  This is to translate the byte storage to real integer value.  This had minor changes to GenotypeTable interface, VCF Export, and MergeIdenticalTaxaPlugin

+++ src/net/maizegenetics/dna/snp/CombineGenotypeTable.java
@@ -819,7 +819,7 @@ public class CombineGenotypeTable implements GenotypeTable {
     }
 
     @Override
-    public byte[] depthForAlleles(int taxon, int site) {
+    public int[] depthForAlleles(int taxon, int site) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
